changed amount to decimal field

+++ buildings/models.py
@@ -29,4 +29,4 @@ class BuildingPicture(TimeStampedModel):
 
 class Penality(TimeStampedModel):
     reason = models.TextField()
-    amount = models.IntegerField()
+    amount = models.DecimalField(decimal_places=2, max_digits=10)
